docs(date-picker): add locale options to stories and docs
