+++ storybook/src/components/DatePicker/DatePicker.docs.mdx
@@ -17,7 +17,7 @@ import * as DatePickerStories from "./DatePicker.stories.tsx";
 
 <Controls />
 
-**Note**: Some props are excluded from the controls panel: `mode`, `onChange`, and `value` are managed by the story’s state wrapper; `defaultMonth`, `isDateDisabled`, `locale`, `maxDate`, and `minDate` have no suitable control widget for their type.
+**Note**: Some props are excluded from the controls panel: `mode`, `onChange`, and `value` are managed by the story’s state wrapper; `defaultMonth`, `isDateDisabled`, `maxDate`, and `minDate` have no suitable control widget for their type.
 
 ## Usage guidelines
 
@@ -40,6 +40,31 @@ In range mode, set `mode="range"`; `value` is a `DateRange` (`{ start, end }`) a
 The first pick sets the start, the second sets the end, and picking before the start begins a new range.
 
 If you set `minDate` or `maxDate`, ensure the initial `value` falls within those bounds — the component does not clamp or validate the incoming value.
+
+#### Localisation
+
+The Date Picker defaults to Dutch.
+Set `locale` and the navigation button label props to use a different language; add `dir="rtl"` for right-to-left scripts such as Arabic.
+In range mode, also localise `rangeStartAccessibleName` and `rangeEndAccessibleName`.
+See the [Language developer guide](/docs/docs-developer-guide-language--docs) for the supported locales.
+
+If your app uses Date Picker in multiple places with the same language, a small wrapper component avoids repeating these props:
+
+```tsx
+const ArabicDatePicker = (props: DatePickerProps) => (
+  <DatePicker
+    dir="rtl"
+    locale="ar-MA"
+    nextMonthButtonLabel="الشهر التالي"
+    nextYearButtonLabel="السنة التالية"
+    previousMonthButtonLabel="الشهر السابق"
+    previousYearButtonLabel="السنة السابقة"
+    rangeEndAccessibleName="تاريخ النهاية"
+    rangeStartAccessibleName="تاريخ البداية"
+    {...props}
+  />
+);
+```
 
 ## Design
